backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
skyfield>=1.49
numpy>=1.26.0
anthropic>=0.34.2
supabase>=2.7.4
stripe>=10.12.0
python-dotenv>=1.0.1
pydantic>=2.9.2
python-multipart>=0.0.12
httpx>=0.27.2
